+++ release/files_common
@@ -340,7 +340,9 @@ selfdrive/ui/*.h
 selfdrive/ui/ui
 selfdrive/ui/text
 selfdrive/ui/spinner
-selfdrive/ui/soundd
+selfdrive/ui/soundd/*.cc
+selfdrive/ui/soundd/*.h
+selfdrive/ui/soundd/soundd
 
 selfdrive/ui/qt/*.cc
 selfdrive/ui/qt/*.h
